com refactor
